FIX: expanding issue in case of mixing DQ and SQ
